refactor: Remove MessageRelay from Value Responder (#3268)
